+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- function to deposit.log resource get Deposit reversal receipt
 ### Fixed 
 - README import on setup file